Performance fixes for appetitive
We can now play the target as early as 150 msec re trial start. This will still often fail due to timing considerations. So, I added a safety check to prevent the program from crashing. The trial just fails and the animal has to repoke. Better approaches can be considered, but we will ignore this for now.