Fix ambiguity in input vs. channel naming
Inputs take precedence when there's an input with the same
name as a channel. We need to be more explicit about this
and ensure that channels have clear naming syntax that does
not clash with input names. Therefore, we now name channels:

   type_code::name

Where type_code is hw_ai, hw_ao, etc.